first complete homology reduction implementation

+++ topologylayer/functional/cohom_cpp/hom.h
@@ -3,4 +3,4 @@
 
 #include "complex.h"
 
-std::vector<torch::Tensor> persistence_forward_hom(SimplicialComplex &X, int MAXDIM);
+std::vector<torch::Tensor> persistence_forward_hom(SimplicialComplex &X, size_t MAXDIM);
